[IMP] removed concept of primary index -> id field is always ID now

+++ fsdb/manager.py
@@ -19,15 +19,6 @@ def dec_check_database_opened(f):
     return wrapper
 
 
-def dec_check_table_exists(f):
-    def wrapper(self, *args, **kwargs):
-        name = args[0] if len(args) > 0 else (kwargs.get('table_name') or kwargs.get('name'))
-        if not self.is_table(name):
-            raise FsdbObjectNotFound('Table with name "{}" does not exist!'.format(name))
-        return f(self, *args, **kwargs)
-    return wrapper
-
-
 class Manager(object):
     # TODO: wait for one transaction to finish before doing next one.
     #  Right not its unsafe to use in threaded environments.
@@ -50,13 +41,19 @@ class Manager(object):
                                 'name': field_name,
                                 'type': field_type,
                                 'default': default_value,  (optional)
+                                'required': False, (optional)
+                                'unique': False,  (optional)
                                 'index': False,  (optional)
-                                'unique_index': False,  (optional)
-                                'primary_index': False,  (required, but in only one field)
                             },
                             ...
                         ],
-                        'records': [{ values }, ...]
+                        'records': [
+                            {
+                                'id': value, (required)
+                                ...
+                            },
+                            ...
+                        ]
                     },
                     ...
                 ]
@@ -76,11 +73,9 @@ class Manager(object):
                     self.create_table(table_config['name'], table_config['fields'])
 
                 for record_values in table_config.get('records', []):
-                    domain = []
-                    for key in record_values:
-                        domain.append((key, '=', record_values[key]))
-                    found_records = self.search_records(table_config['name'], domain, limit=1)
-                    if len(found_records) == 0:
+                    rid = record_values['id']
+                    found = self.browse_records(table_config['name'], rid)
+                    if not found:
                         self.create_record(table_config['name'], record_values)
 
             self.close_database()
@@ -122,44 +117,45 @@ class Manager(object):
         return True if (name in self.database.tables) else False
 
     @dec_check_database_opened
+    def get_table(self, name):
+        if not self.is_table(name):
+            raise FsdbObjectNotFound('Table with name "{}" does not exist!'.format(name))
+        return self.database.tables[name]
+
+    @dec_check_database_opened
     def create_table(self, name, fields):
         return Table.create(self.database, name, fields)
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def delete_table(self, name):
-        self.database.tables[name].delete()
-        del(self.database.tables[name])
+        if self.is_table(name):
+            self.database.tables[name].delete()
+            del(self.database.tables[name])
 
     # Record
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def create_record(self, table_name, values):
-        table = self.database.tables[table_name]
+        table = self.get_table(table_name)
         return Record.create(table, values)
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def write_records(self, table_name, values, domain=None):
         records = self.search_records(table_name, domain)
         map(lambda rec: rec.write(values), records)
         return records
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def browse_records(self, table_name, ids):
-        table = self.database.tables[table_name]
+        table = self.get_table(table_name)
         return table.browse_records(ids)
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def search_records(self, table_name, domain=None, limit=None):
-        table = self.database.tables[table_name]
+        table = self.get_table(table_name)
         return table.search_records(domain=domain, limit=limit)
 
     @dec_check_database_opened
-    @dec_check_table_exists
     def delete_records(self, table_name, domain=None):
         records = self.search_records(table_name, domain)
         map(lambda x: x.delete(), records)
